pyproject.toml
[project]
authors = [{name = "Geoffrey Hutchison", email = "geoff.hutchison@gmail.com"}]
dependencies = ["mace-torch>=0.3.15,<0.4"]
name = "avogadro-mace"
requires-python = ">= 3.11"
version = "0.1.0"
license = "BSD-3-Clause"
description = "Energy model using MACE ML potentials"

[project.scripts]
avogadro-mace = "avogadro_mace:main"

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[tool.hatch.build.targets.wheel]
packages = ["src/avogadro_mace"]

[tool.avogadro]
minimum-avogadro-version = "1.103"

[[tool.avogadro.energy-models]]
identifier = "MACE-MP-0"
model-name = "MACE-MP-0"
description = "Calculate MACE energies and gradients for inorganic crystals"
input-format = "cjson"
protocol = "binary-v1"
support.gradients = true
support.elements = "1-89"
support.unit-cell = true
support.ions = true
support.radicals = true

[[tool.avogadro.energy-models]]
identifier = "MACE-OFF23"
model-name = "MACE-OFF23"
description = "Calculate MACE energies and gradients for organic molecules and crystals"
input-format = "cjson"
protocol = "binary-v1"
support.gradients = true
support.elements = "1, 6, 7, 8, 9, 15, 16, 17, 35, 53"
support.unit-cell = true
support.ions = false
support.radicals = false

[tool.pixi.workspace]
channels = ["conda-forge"]
platforms = ["win-64", "linux-64", "linux-aarch64", "osx-arm64"]

[tool.pixi.pypi-dependencies]
avogadro_mace = { path = ".", editable = true }
mace-torch = ">=0.3.15,<0.4"

[tool.pixi.tasks]
